Created the AbstractPathMappingRepository to remove code duplication
